gdbus: Fix Bluetooth LE mouse "Connected" prop not updating
When a Bluetooth LE mouse with the Battery1 interface gets disconnected,
the "Connected" property is queued for update and then the Battery1
interface gets removed.

But as remove_interface() sets data->pending_prop to FALSE without
checking on the other interfaces, the queued property changes on the
other interfaces are just ignored.

They would eventually be flushed when other properties of that device
eventually changed.

We fix this by only setting the data->pending_prop value to FALSE when
we're certain to have processed all the interfaces, and when we've
checked other interfaces don't have any pending props.